remove cfg_and_block_env method
