+++ crates/node-builder/src/evm.rs
@@ -42,3 +42,29 @@ impl EvmEnvConfig for EthEvmConfig {
         cfg_env.perf_analyse_created_bytecodes = AnalysisKind::Analyse;
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use reth_primitives::revm_primitives::BlockEnv;
+
+    #[test]
+    #[ignore]
+    fn test_fill_cfg_and_block_env() {
+        let mut cfg_env = CfgEnv::default();
+        let mut block_env = BlockEnv::default();
+        let header = Header::default();
+        let chain_spec = ChainSpec::default();
+        let total_difficulty = U256::ZERO;
+
+        EthEvmConfig::fill_cfg_and_block_env(
+            &mut cfg_env,
+            &mut block_env,
+            &chain_spec,
+            &header,
+            total_difficulty,
+        );
+
+        assert_eq!(cfg_env.chain_id, chain_spec.chain().id());
+    }
+}
